Move notification action functions from shared/ to whats-new-popup.js
