+++ shared/notifications/index.js
@@ -1,6 +1,3 @@
-import { ETH_SWAPS_TOKEN_OBJECT } from '../constants/swaps';
-import { BUILD_QUOTE_ROUTE } from '../../ui/app/helpers/constants/routes';
-
 // Messages and descriptions for these locale keys are in app/_locales/en/messages.json
 export const UI_NOTIFICATIONS = {
   1: {
@@ -26,34 +23,3 @@ export const UI_NOTIFICATIONS = {
     actionText: 'notificationsActionText3',
   },
 };
-
-export function notifcationActionFunctions(metricsEvent) {
-  const actionFunctions = {
-    1: () => {
-      metricsEvent({
-        event: 'Swaps Opened',
-        properties: { source: 'Main View', active_currency: 'ETH' },
-        category: 'swaps',
-      });
-      global.platform.openExtensionInBrowser(
-        BUILD_QUOTE_ROUTE,
-        `fromAddress=${ETH_SWAPS_TOKEN_OBJECT.address}`,
-      );
-    },
-    2: () => {
-      global.platform.openTab({
-        url: 'https://metamask.io/download.html',
-      });
-    },
-    3: () => {
-      global.platform.openTab({
-        url:
-          'https://survey.alchemer.com/s3/6173069/MetaMask-Extension-NPS-January-2021',
-      });
-    },
-  };
-
-  return (id) => {
-    return actionFunctions[id];
-  };
-}
